requirements.txt
numpy
pandas
plotly.express
streamlit
openpyxl
